Restructure ResponseEntity features for StatusController

+++ Spring/ParkingStatus/src/main/java/com/ParkingStatus/ParkingStatus/DataAccessService/Status/StatusDataAccessService.java
@@ -123,13 +123,17 @@ public class StatusDataAccessService {
     public int removeStatus(int id){
         //remove outer dependencies first
 
-        removeOuterDependencies(id);
+        if(selectStatusById(id) != null){
+            removeOuterDependencies(id);
 
-        String removeSql = "DELETE FROM status WHERE " +
-                "statusId = " + id +";";
+            String removeSql = "DELETE FROM status WHERE " +
+                    "statusId = " + id +";";
 
-        jdbcTemplate.update(removeSql);
-        return id;
+            jdbcTemplate.update(removeSql);
+            return id;
+        }
+
+        return -1;
     }
